rpi-requirements.txt
numpy==2.2.6
opencv-contrib-python==4.12.0.88
pillow==12.1.0
setuptools==80.9.0
spidev==3.8
